created a UserTable.tsx component inside of the Users folder that will dynamically route the user table inside of the page.tsx component when routed inside of it.
